The code is reduced to a modular type
