+++ index.html
@@ -231,5 +231,5 @@
   </template>
 
 </body>
-<script src="/js/main.js"></script>
+<script src="/js/main.js" type="module"></script>
 </html>
