+++ js/data.js
@@ -1,0 +1,41 @@
+const NAMES = [
+  'Иван',
+  'Анастасия',
+  'Мария',
+  'Ибрагим',
+  'Виктор',
+  'Алия',
+  'Кекс',
+  'Василий',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const DESCRIPTION = [
+  'Смейтесь как только умеете, любите столько, сколько живете.',
+  'Жизнь похожа на фотокамеру: вам просто нужно смотреть на нее с улыбкой.',
+  'Жизнь — это всего лишь серия крошечных чудес, поэтому обратите внимание на них.',
+  'Я пыталась заниматься йогой, но в позе лотоса уснула.',
+  'Ты любишь всех, а любить всех — значит не любить никого.',
+  'Циник — это человек, который всему знает цену и ничего не ценит.',
+  'Опытом люди называют свои ошибки.',
+  'В жизни бывают только две настоящие трагедии: одна — когда не получаешь того, чего хочешь, а вторая — когда получаешь.',
+  'Кто эти люди? Бойцы? Нисколько.Может быть, партизаны? Нет.',
+  'Жди меня, и я вернусь. Только очень жди...',
+  'В одно окно смотрели двое. Один увидел дождь и грязь. Другой — листвы зелёной вязь,весну и небо голубое.В одно окно смотрели двое',
+  'Чтоб мудро жизнь прожить, знать надобно немало, Два важных правила запомни для начала: Ты лучше голодай, чем что попало есть,И лучше будь один, чем вместе с кем попало.',
+];
+
+const POST_QUANTITY = 25;
+const ID_NUMBER = 1000;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const COMMENTS_ID = [];
+
+export {NAMES, COMMENTS, DESCRIPTION, POST_QUANTITY, ID_NUMBER, MAX_LIKES, MIN_LIKES, COMMENTS_ID};
